Import the Box Android SDK as a sub-project
The normal Box SDK has both a Java and Android library that must be built
together.  It also has a library which conflicts with Google Drive,
jackson-core.  So, the distribution SDK projects are built externally and then
imported into this box-lib project for use by PasswdSafe Sync.

The Java SDK is built first via Ant.  The build file is updated to force source
and target use of JDK 1.6.  The resulting BoxJavaLibraryV2.jar is copied into
box-lib.

The Android SDK is built next using Eclipse.  The resulting
bin/boxandroidlibraryv2.jar and all of the project's resources are copied into
box-lib.